abstract teleop

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/ClawSlide.java
@@ -11,9 +11,9 @@ import org.firstinspires.ftc.teamcode.action.MotorPairAction;
 import org.firstinspires.ftc.teamcode.action.TimedUpdateAction;
 
 public class ClawSlide {
-    private static final double ROTATE_POWER = 0.4;
-    private static final double LIFT_POWER1 = 0.8;
-    private static final double LIFT_POWER2 = 0.8;
+    private static final float ROTATE_POWER = 0.4f;
+    private static final float LIFT_POWER1 = 0.8f;
+    private static final float LIFT_POWER2 = 0.8f;
 
     public static final int ROTATE_MAX_POSITION = 1090;
     private static final double ROTATE_ANGLE_RATIO = 90.0 / 990;
